Source: https://github.com/msupply-foundation/open-msupply/commit/081c21bc7982a594dcc8b32c8f56076badcab081

+++ server/service/src/barcode.rs
@@ -16,13 +16,13 @@ pub struct InsertResult {
     pub id: String,
     pub gtin: String,
     pub item_id: String,
-    pub pack_size: Option<i32>,
+    pub pack_size: Option<f64>,
 }
 
 pub struct BarcodeInput {
     pub gtin: String,
     pub item_id: String,
-    pub pack_size: Option<i32>,
+    pub pack_size: Option<f64>,
 }
 
 #[derive(Debug, PartialEq)]
